ROX-12263: fix authz plugin removal migrator crash on recent DB (#2742)

+++ migrator/migrations/m_107_to_m_108_remove_auth_plugin/migration.go
@@ -72,6 +72,10 @@ func migratePS(db *gorocksdb.DB) error {
 
 func deleteAuthPluginBucket(db *bbolt.DB) error {
 	return db.Update(func(tx *bbolt.Tx) error {
+		if b := tx.Bucket(authPluginBucket); b == nil {
+			// nothing to be done if the bucket doesn't exist
+			return nil
+		}
 		return tx.DeleteBucket(authPluginBucket)
 	})
 }
